+++ NEWS.md
@@ -5,9 +5,10 @@
 * refactor: Optimize runtime of setting row roles.
 * refactor: Optimize runtime of marshalling.
 * refactor: Optimize runtime of `Task$col_info`
+* fix: `Task$cbind()` now works with non-standard primary keys
+  for `data.frames` (#961).
 * fix: Triggering of fallback learner now has log-level "info"
   instead of "debug" (#972)
-
 
 # mlr3 0.20.2